refactor: chatStu 파일 구조 재정리(#47) - routes 폴더 안에 ChatStu 폴더 만들어서 ChatStu.jsx, ChatStu.css 욺겼다 - hooks 폴더 안에 useChatStu.js 파일을 옮겼다 - App.jsx 파일에 경로 수정 - 메시지 보낼 때 e.preventDefault 추가
